refactor common JSON test utils
