+++ common/src/test/java/org/example/age/common/testing/JsonTestingTest.java
@@ -1,6 +1,5 @@
 package org.example.age.common.testing;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
@@ -11,7 +10,7 @@ import org.junit.jupiter.api.Test;
 public final class JsonTestingTest {
 
     @Test
-    public void serializeThenDeserialize_String() throws IOException {
+    public void serializeThenDeserialize() throws IOException {
         JsonTesting.serializeThenDeserialize("test", String.class);
     }
 
@@ -19,13 +18,6 @@ public final class JsonTestingTest {
     public void serializeThenDeserialize_BadValue() {
         assertThatThrownBy(() -> JsonTesting.serializeThenDeserialize(BadValue.fromString("value"), BadValue.class))
                 .isInstanceOf(AssertionError.class);
-    }
-
-    @Test
-    public void serializeThenDeserialize_TestUtils() {
-        String json = JsonTesting.serialize("test");
-        String value = JsonTesting.deserialize(json, String.class);
-        assertThat(value).isEqualTo("test");
     }
 
     /** Value type that serializes incorrectly. */
